runtime: add evaluator registry for run
