+++ src/onnx2c/runtime/__init__.py
@@ -1,0 +1,1 @@
+"""Runtime helpers for evaluating ONNX graphs."""
